split the code - moved parts of code into _TMC_2209_comm - moved parts of code into _TMC_2209_move - renamed modules that should not be imported externally with _ prefix - renamed read register functions to snake_case
