+++ src/TMC_2209/TMC_2209_StepperDriver.py
@@ -5,48 +5,20 @@
 #pylint: disable=too-many-public-methods
 #pylint: disable=too-many-branches
 #pylint: disable=too-many-instance-attributes
+#pylint: disable=import-outside-toplevel
 """
 TMC_2209 stepper driver module
 """
 
 import time
-from enum import Enum
-import math
 import statistics
-import threading
 from RPi import GPIO
-from .TMC_2209_uart import TMC_UART as tmc_uart
-from .TMC_2209_logger import TMC_logger as tmc_logger, Loglevel
-from . import TMC_2209_reg as tmc_reg
-from . import TMC_2209_math as tmc_math
+from ._TMC_2209_uart import TMC_UART as tmc_uart
+from ._TMC_2209_logger import TMC_logger, Loglevel
+from ._TMC_2209_move import MovementAbsRel, MovementPhase, StopMode
+from . import _TMC_2209_reg as tmc_reg
+from . import _TMC_2209_math as tmc_math
 
-
-
-class Direction(Enum):
-    """movement direction of the motor"""
-    CCW = 0
-    CW = 1
-
-
-class MovementAbsRel(Enum):
-    """movement absolute or relative"""
-    ABSOLUTE = 0
-    RELATIVE = 1
-
-
-class MovementPhase(Enum):
-    """movement phase"""
-    STANDSTILL = 0
-    ACCELERATING = 1
-    MAXSPEED = 2
-    DECELERATING = 3
-
-
-class StopMode(Enum):
-    """stopmode"""
-    NO = 0
-    SOFTSTOP = 1
-    HARDSTOP = 2
 
 
 
@@ -58,6 +30,25 @@ class TMC_2209:
     1. change setting in the TMC-driver via UART
     2. move the motor via STEP/DIR pins
     """
+
+    from ._TMC_2209_comm import (
+        read_drv_status, read_gconf, read_gstat, clear_gstat, read_ioin, read_chopconf,
+        get_direction_reg, set_direction_reg, get_iscale_analog, set_iscale_analog,get_vsense,
+        set_vsense, get_internal_rsense, set_internal_rsense, set_irun_ihold, set_pdn_disable,
+        set_current, get_spreadcycle, set_spreadcycle, get_interpolation, set_interpolation,
+        read_microstepping_resolution, get_microstepping_resolution, set_microstepping_resolution,
+        set_mstep_resolution_reg_select, get_interface_transmission_counter, get_tstep, set_vactual,
+        get_stallguard_result, set_stallguard_threshold, set_coolstep_threshold,
+        get_microstep_counter, get_microstep_counter_in_steps
+    )
+
+    from ._TMC_2209_move import (
+        set_max_speed, set_max_speed_fullstep, get_max_speed, set_acceleration,
+        set_acceleration_fullstep, get_acceleration, stop, get_movement_phase,
+        run_to_position_steps, run_to_position_revolutions, run_to_position_steps_threaded,
+        run_to_position_revolutions_threaded, wait_for_movement_finished_threaded,
+        run, distance_to_go, compute_new_speed, run_speed, make_a_step
+    )
 
     tmc_uart = None
     tmc_logger = None
@@ -103,11 +94,11 @@ class TMC_2209:
 
 
     def __init__(self, pin_en, pin_step=-1, pin_dir=-1, baudrate=115200, serialport="/dev/serial0",
-                 driver_address=0, no_uart=False, gpio_mode=GPIO.BCM, loglevel = None):
+                 driver_address=0, gpio_mode=GPIO.BCM, loglevel = None):
         """
         constructor
         """
-        self.init(pin_en, pin_step, pin_dir, baudrate, serialport, driver_address, no_uart,
+        self.init(pin_en, pin_step, pin_dir, baudrate, serialport, driver_address,
                   gpio_mode, loglevel)
 
 
@@ -117,11 +108,13 @@ class TMC_2209:
         destructor
         """
         self.deinit()
+        del self.tmc_uart
+        del self.tmc_logger
 
 
 
     def init(self, pin_en, pin_step=-1, pin_dir=-1, baudrate=115200, serialport="/dev/serial0",
-    driver_address=0, no_uart=False, gpio_mode=GPIO.BCM, loglevel = None):
+    driver_address=0, gpio_mode=GPIO.BCM, loglevel = None):
         """
         init function
         
@@ -136,38 +129,35 @@ class TMC_2209:
                 gpio_mode (enum): numbering system for the GPIO pins
                 loglevel (enum): level for which to log
         """
-        self.tmc_uart = tmc_uart(serialport, baudrate, driver_address)
-        self.tmc_logger = tmc_logger(Loglevel.INFO, "TMC2209 "+str(driver_address))
+        self.tmc_logger = TMC_logger(loglevel, "TMC2209 "+str(driver_address))
+        self.tmc_uart = tmc_uart(self.tmc_logger, serialport, baudrate, driver_address)
 
-        if loglevel is not None:
-            self._loglevel = loglevel
 
-        self.tmc_logger.log("Init", Loglevel.INFO.value)
+        self.tmc_logger.log("Init", Loglevel.INFO)
         GPIO.setwarnings(False)
         GPIO.setmode(gpio_mode)
 
-        self.tmc_logger.log("EN Pin: " + str(pin_en), Loglevel.DEBUG.value)
+        self.tmc_logger.log("EN Pin: " + str(pin_en), Loglevel.DEBUG)
         self._pin_en = pin_en
         GPIO.setup(self._pin_en, GPIO.OUT, initial=GPIO.HIGH)
 
-        self.tmc_logger.log("STEP Pin: " + str(pin_step), Loglevel.DEBUG.value)
+        self.tmc_logger.log("STEP Pin: " + str(pin_step), Loglevel.DEBUG)
         if pin_step != -1:
             self._pin_step = pin_step
             GPIO.setup(self._pin_step, GPIO.OUT, initial=GPIO.LOW)
 
-        self.tmc_logger.log("DIR Pin: " + str(pin_dir), Loglevel.DEBUG.value)
+        self.tmc_logger.log("DIR Pin: " + str(pin_dir), Loglevel.DEBUG)
         if pin_dir != -1:
             self._pin_dir = pin_dir
             GPIO.setup(self._pin_dir, GPIO.OUT, initial=self._direction)
 
-        self.tmc_logger.log("GPIO Init finished", Loglevel.INFO.value)
+        self.tmc_logger.log("GPIO Init finished", Loglevel.INFO)
 
-        if not no_uart:
-            self.read_steps_per_rev()
-            self.clearGSTAT()
+        self.read_steps_per_rev()
+        self.clear_gstat()
 
         self.tmc_uart.flush_serial_buffer()
-        self.tmc_logger.log("Init finished", Loglevel.INFO.value)
+        self.tmc_logger.log("Init finished", Loglevel.INFO)
 
 
 
@@ -176,7 +166,7 @@ class TMC_2209:
         deinit function
         """
         if self._deinit_finished is False:
-            self.tmc_logger.log("Deinit", Loglevel.INFO.value)
+            self.tmc_logger.log("Deinit", Loglevel.INFO)
 
             self.set_motor_enabled(False)
 
@@ -191,10 +181,10 @@ class TMC_2209:
                 GPIO.remove_event_detect(self._pin_stallguard)
                 GPIO.cleanup(self._pin_stallguard)
 
-            self.tmc_logger.log("Deinit finished", Loglevel.INFO.value)
+            self.tmc_logger.log("Deinit finished", Loglevel.INFO)
             self._deinit_finished= True
         else:
-            self.tmc_logger.log("Deinit already finished", Loglevel.INFO.value)
+            self.tmc_logger.log("Deinit already finished", Loglevel.INFO)
 
 
 
@@ -218,215 +208,6 @@ class TMC_2209:
 
 
 
-    def readDRVSTATUS(self):
-        """
-        read the register Adress "DRV_STATUS" and prints all current setting
-
-            Returns:
-                drvstatus (int): 32bit DRV_STATUS Register
-
-        """
-        self.tmc_logger.log("---")
-        self.tmc_logger.log("DRIVER STATUS:")
-        drvstatus =self.tmc_uart.read_int(tmc_reg.DRVSTATUS)
-        self.tmc_logger.log(bin(drvstatus), Loglevel.INFO.value)
-        if drvstatus & tmc_reg.stst:
-            self.tmc_logger.log("Info: motor is standing still")
-        else:
-            self.tmc_logger.log("Info: motor is running")
-
-        if drvstatus & tmc_reg.stealth:
-            self.tmc_logger.log("Info: motor is running on StealthChop")
-        else:
-            self.tmc_logger.log("Info: motor is running on SpreadCycle")
-
-        cs_actual = drvstatus & tmc_reg.cs_actual
-        cs_actual = cs_actual >> 16
-        self.tmc_logger.log("CS actual: "+str(cs_actual))
-
-        if drvstatus & tmc_reg.olb:
-            self.tmc_logger.log("Warning: Open load detected on phase B")
-
-        if drvstatus & tmc_reg.ola:
-            self.tmc_logger.log("Warning: Open load detected on phase A")
-
-        if drvstatus & tmc_reg.s2vsb:
-            self.tmc_logger.log("""Error: Short on low-side MOSFET detected on phase B.
-                     The driver becomes disabled""")
-
-        if drvstatus & tmc_reg.s2vsa:
-            self.tmc_logger.log("""Error: Short on low-side MOSFET detected on phase A.
-                     The driver becomes disabled""")
-
-        if drvstatus & tmc_reg.s2gb:
-            self.tmc_logger.log("""Error: Short to GND detected on phase B.
-                                The driver becomes disabled.""")
-
-        if drvstatus & tmc_reg.s2ga:
-            self.tmc_logger.log("""Error: Short to GND detected on phase A.
-                                The driver becomes disabled.""")
-
-        if drvstatus & tmc_reg.ot:
-            self.tmc_logger.log("Error: Driver Overheating!")
-
-        if drvstatus & tmc_reg.otpw:
-            self.tmc_logger.log("Warning: Driver Overheating Prewarning!")
-
-        self.tmc_logger.log("---")
-        return drvstatus
-
-
-
-    def readGCONF(self):
-        """
-        read the register Adress "GCONF" and prints all current setting
-
-            Returns:
-                gconf (int): 10bit GCONF Register
-        """
-        self.tmc_logger.log("---")
-        self.tmc_logger.log("GENERAL CONFIG")
-        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
-        self.tmc_logger.log(bin(gconf), Loglevel.INFO.value)
-
-        if gconf & tmc_reg.i_scale_analog:
-            self.tmc_logger.log("Driver is using voltage supplied to VREF as current reference")
-        else:
-            self.tmc_logger.log("Driver is using internal reference derived from 5VOUT")
-        if gconf & tmc_reg.internal_rsense:
-            self.tmc_logger.log("""Internal sense resistors.
-                                Use current supplied into VREF as reference.""")
-            self.tmc_logger.log("VREF pin internally is driven to GND in this mode.")
-            self.tmc_logger.log("This will most likely destroy your driver!!!")
-            raise SystemExit
-        self.tmc_logger.log("Operation with external sense resistors")
-        if gconf & tmc_reg.en_spreadcycle:
-            self.tmc_logger.log("SpreadCycle mode enabled")
-        else:
-            self.tmc_logger.log("StealthChop PWM mode enabled")
-        if gconf & tmc_reg.shaft:
-            self.tmc_logger.log("Inverse motor direction")
-        else:
-            self.tmc_logger.log("normal motor direction")
-        if gconf & tmc_reg.index_otpw:
-            self.tmc_logger.log("INDEX pin outputs overtemperature prewarning flag")
-        else:
-            self.tmc_logger.log("INDEX shows the first microstep position of sequencer")
-        if gconf & tmc_reg.index_step:
-            self.tmc_logger.log("INDEX output shows step pulses from internal pulse generator")
-        else:
-            self.tmc_logger.log("INDEX output as selected by index_otpw")
-        if gconf & tmc_reg.mstep_reg_select:
-            self.tmc_logger.log("Microstep resolution selected by MSTEP register")
-        else:
-            self.tmc_logger.log("Microstep resolution selected by pins MS1, MS2")
-
-        self.tmc_logger.log("---")
-        return gconf
-
-
-
-    def readGSTAT(self):
-        """
-        read the register Adress "GSTAT" and prints all current setting
-
-            Returns:
-                gstat (int): 3bit GSTAT Register
-        """
-        self.tmc_logger.log("---")
-        self.tmc_logger.log("GSTAT")
-        gstat = self.tmc_uart.read_int(tmc_reg.GSTAT)
-        self.tmc_logger.log(bin(gstat), Loglevel.INFO.value)
-        if gstat & tmc_reg.reset:
-            self.tmc_logger.log("The Driver has been reset since the last read access to GSTAT")
-        if gstat & tmc_reg.drv_err:
-            self.tmc_logger.log("""The driver has been shut down due to overtemperature or
-                     short circuit detection since the last read access""")
-        if gstat & tmc_reg.uv_cp:
-            self.tmc_logger.log("""Undervoltage on the charge pump.
-                                The driver is disabled in this case""")
-        self.tmc_logger.log("---")
-        return gstat
-
-
-
-    def clearGSTAT(self):
-        """
-        read the register Adress "GSTAT" and prints all current setting
-        """
-        self.tmc_logger.log("clearing GSTAT", Loglevel.INFO.value)
-        gstat = self.tmc_uart.read_int(tmc_reg.GSTAT)
-
-        gstat = self.tmc_uart.set_bit(gstat, tmc_reg.reset)
-        gstat = self.tmc_uart.set_bit(gstat, tmc_reg.drv_err)
-
-        self.tmc_uart.write_reg_check(tmc_reg.GSTAT, gstat)
-
-
-
-    def readIOIN(self):
-        """
-        read the register Adress "IOIN" and prints all current setting
-
-            Returns:
-                ioin (int): 10+8bit IOIN Register
-        """
-        self.tmc_logger.log("---")
-        self.tmc_logger.log("INPUTS")
-        ioin = self.tmc_uart.read_int(tmc_reg.IOIN)
-        self.tmc_logger.log(bin(ioin), Loglevel.INFO.value)
-        if ioin & tmc_reg.io_spread:
-            self.tmc_logger.log("spread is high")
-        else:
-            self.tmc_logger.log("spread is low")
-
-        if ioin & tmc_reg.io_dir:
-            self.tmc_logger.log("dir is high")
-        else:
-            self.tmc_logger.log("dir is low")
-
-        if ioin & tmc_reg.io_step:
-            self.tmc_logger.log("step is high")
-        else:
-            self.tmc_logger.log("step is low")
-
-        if ioin & tmc_reg.io_enn:
-            self.tmc_logger.log("en is high")
-        else:
-            self.tmc_logger.log("en is low")
-
-        self.tmc_logger.log("---")
-        return ioin
-
-
-
-    def readCHOPCONF(self):
-        """
-        read the register Adress "CHOPCONF" and prints all current setting
-
-            Returns:
-                chopconf (int): 3bit CHOPCONF Register
-        """
-        self.tmc_logger.log("---")
-        self.tmc_logger.log("CHOPPER CONTROL")
-        chopconf = self.tmc_uart.read_int(tmc_reg.CHOPCONF)
-        self.tmc_logger.log(bin(chopconf), Loglevel.INFO.value)
-
-        self.tmc_logger.log("native "+str(self.get_microstepping_resolution())+" microstep setting")
-
-        if chopconf & tmc_reg.intpol:
-            self.tmc_logger.log("interpolation to 256 microsteps")
-
-        if chopconf & tmc_reg.vsense:
-            self.tmc_logger.log("1: High sensitivity, low sense resistor voltage")
-        else:
-            self.tmc_logger.log("0: Low sensitivity, high sense resistor voltage")
-
-        self.tmc_logger.log("---")
-        return chopconf
-
-
-
     def set_motor_enabled(self, en):
         """
         enables or disables the motor current output
@@ -435,7 +216,7 @@ class TMC_2209:
                 en (bool): whether the motor current output should be enabled
         """
         GPIO.output(self._pin_en, not en)
-        self.tmc_logger.log(f"Motor output active: {en}", Loglevel.INFO.value)
+        self.tmc_logger.log(f"Motor output active: {en}", Loglevel.INFO)
 
 
 
@@ -458,13 +239,13 @@ class TMC_2209:
         if speed_rpm is None:
             speed_rpm = tmc_math.steps_to_rps(self._max_speed_homing, self._steps_per_rev)*60
 
-        self.tmc_logger.log("---", Loglevel.INFO.value)
-        self.tmc_logger.log("homing", Loglevel.INFO.value)
+        self.tmc_logger.log("---", Loglevel.INFO)
+        self.tmc_logger.log("homing", Loglevel.INFO)
 
         # StallGuard only works with StealthChop
         self.set_spreadcycle(0)
 
-        self.tmc_logger.log("Stallguard threshold:"+str(self._sg_threshold), Loglevel.DEBUG.value)
+        self.tmc_logger.log("Stallguard threshold:"+str(self._sg_threshold), Loglevel.DEBUG)
 
         self.set_stallguard_callback(diag_pin, self._sg_threshold, self.stop,
                                      0.5*tmc_math.rps_to_steps(speed_rpm/60, self._steps_per_rev))
@@ -472,13 +253,13 @@ class TMC_2209:
         homing_failed = self.set_vactual_rpm(speed_rpm, revolutions=revolutions)
 
         if homing_failed:
-            self.tmc_logger.log("homing failed", Loglevel.INFO.value)
+            self.tmc_logger.log("homing failed", Loglevel.INFO)
         else:
-            self.tmc_logger.log("homing successful",Loglevel.INFO.value)
+            self.tmc_logger.log("homing successful",Loglevel.INFO)
 
         self._current_pos = 0
 
-        self.tmc_logger.log("---", Loglevel.INFO.value)
+        self.tmc_logger.log("---", Loglevel.INFO)
         return not homing_failed
 
 
@@ -498,10 +279,10 @@ class TMC_2209:
         if threshold is not None:
             self._sg_threshold = threshold
 
-        self.tmc_logger.log("---", Loglevel.INFO.value)
-        self.tmc_logger.log("homing", Loglevel.INFO.value)
+        self.tmc_logger.log("---", Loglevel.INFO)
+        self.tmc_logger.log("homing", Loglevel.INFO)
 
-        self.tmc_logger.log("Stallguard threshold:"+str(self._sg_threshold), Loglevel.DEBUG.value)
+        self.tmc_logger.log("Stallguard threshold:"+str(self._sg_threshold), Loglevel.DEBUG)
 
         self.set_direction_pin(revolutions > 0)
 
@@ -514,7 +295,8 @@ class TMC_2209:
         self._n = 0
         self.set_acceleration(10000)
         self.set_max_speed(self._max_speed_homing)
-        coolstep_thres = tmc_math.steps_to_tstep(self._max_speed_homing*0.5, self._msres)
+        coolstep_thres = tmc_math.steps_to_tstep(self._max_speed_homing*0.5,
+                                                 self.get_microstepping_resolution())
         self.set_coolstep_threshold(coolstep_thres)
         self.compute_new_speed()
 
@@ -533,16 +315,16 @@ class TMC_2209:
                         break
 
         if step_counter<self._steps_per_rev:
-            self.tmc_logger.log("homing successful",Loglevel.INFO.value)
-            self.tmc_logger.log("Stepcounter: "+str(step_counter),Loglevel.DEBUG.value)
-            self.tmc_logger.log(str(sg_results),Loglevel.DEBUG.value)
+            self.tmc_logger.log("homing successful",Loglevel.INFO)
+            self.tmc_logger.log("Stepcounter: "+str(step_counter),Loglevel.DEBUG)
+            self.tmc_logger.log(str(sg_results),Loglevel.DEBUG)
             self._current_pos = 0
         else:
-            self.tmc_logger.log("homing failed", Loglevel.INFO.value)
-            self.tmc_logger.log("Stepcounter: "+str(step_counter), Loglevel.DEBUG.value)
-            self.tmc_logger.log(str(sg_results),Loglevel.DEBUG.value)
+            self.tmc_logger.log("homing failed", Loglevel.INFO)
+            self.tmc_logger.log("Stepcounter: "+str(step_counter), Loglevel.DEBUG)
+            self.tmc_logger.log(str(sg_results),Loglevel.DEBUG)
 
-        self.tmc_logger.log("---", Loglevel.INFO.value)
+        self.tmc_logger.log("---", Loglevel.INFO)
 
 
 
@@ -589,381 +371,6 @@ class TMC_2209:
 
 
 
-    def get_direction_reg(self):
-        """
-        returns the motor shaft direction: 0 = CCW; 1 = CW
-
-            Returns:
-                direction (bool): motor shaft direction: 0 = CCW; 1 = CW
-        """
-        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
-        return gconf & tmc_reg.shaft
-
-
-
-    def set_direction_reg(self, direction):
-        """
-        sets the motor shaft direction to the given value: 0 = CCW; 1 = CW
-
-            Parameters:
-                direction (bool): motor shaft direction: 0 = CCW; 1 = CW
-        """
-        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
-        if direction:
-            self.tmc_logger.log("write inverse motor direction", Loglevel.INFO.value)
-            gconf = self.tmc_uart.set_bit(gconf, tmc_reg.shaft)
-        else:
-            self.tmc_logger.log("write normal motor direction", Loglevel.INFO.value)
-            gconf = self.tmc_uart.clear_bit(gconf, tmc_reg.shaft)
-        self.tmc_uart.write_reg_check(tmc_reg.GCONF, gconf)
-
-
-
-    def get_iscale_analog(self):
-        """
-        return whether Vref (1) or 5V (0) is used for current scale
-
-            Returns:
-                en (bool): whether Vref (1) or 5V (0) is used for current scale
-        """
-        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
-        return gconf & tmc_reg.i_scale_analog
-
-
-
-    def set_iscale_analog(self,en):
-        """
-        sets Vref (1) or 5V (0) for current scale
-
-            Parameters:
-                en (bool): whether Vref (1) or 5V (0) is used for current scale
-        """
-        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
-        if en:
-            self.tmc_logger.log("activated Vref for current scale", Loglevel.INFO.value)
-            gconf = self.tmc_uart.set_bit(gconf, tmc_reg.i_scale_analog)
-        else:
-            self.tmc_logger.log("activated 5V-out for current scale", Loglevel.INFO.value)
-            gconf = self.tmc_uart.clear_bit(gconf, tmc_reg.i_scale_analog)
-        self.tmc_uart.write_reg_check(tmc_reg.GCONF, gconf)
-
-
-
-    def getvsense(self):
-        """
-        returns which sense resistor voltage is used for current scaling
-        0: Low sensitivity, high sense resistor voltage
-        1: High sensitivity, low sense resistor voltage
-
-            Returns:
-                en (bool): whether high sensitivity should is used
-        """
-        chopconf = self.tmc_uart.read_int(tmc_reg.CHOPCONF)
-        return chopconf & tmc_reg.vsense
-
-
-
-    def setvsense(self,en):
-        """
-        sets which sense resistor voltage is used for current scaling
-        0: Low sensitivity, high sense resistor voltage
-        1: High sensitivity, low sense resistor voltage
-
-            Parameters:
-                en (bool): whether high sensitivity should be used
-        """
-        chopconf = self.tmc_uart.read_int(tmc_reg.CHOPCONF)
-        if en:
-            self.tmc_logger.log("activated High sensitivity, low sense resistor voltage",
-                                Loglevel.INFO.value)
-            chopconf = self.tmc_uart.set_bit(chopconf, tmc_reg.vsense)
-        else:
-            self.tmc_logger.log("activated Low sensitivity, high sense resistor voltage",
-                                Loglevel.INFO.value)
-            chopconf = self.tmc_uart.clear_bit(chopconf, tmc_reg.vsense)
-        self.tmc_uart.write_reg_check(tmc_reg.CHOPCONF, chopconf)
-
-
-
-    def get_internal_rsense(self):
-        """
-        returns which sense resistor voltage is used for current scaling
-        0: Operation with external sense resistors
-        1: Internal sense resistors. Use current supplied into
-        VREF as reference for internal sense resistor. VREF
-        pin internally is driven to GND in this mode.
-
-            Returns:
-                en (bool): which sense resistor voltage is used
-        """
-        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
-        return gconf & tmc_reg.internal_rsense
-
-
-
-    def set_internal_rsense(self,en):
-        """
-        sets which sense resistor voltage is used for current scaling
-        0: Operation with external sense resistors
-        1: Internal sense resistors. Use current supplied into
-        VREF as reference for internal sense resistor. VREF
-        pin internally is driven to GND in this mode.
-
-            Parameters:
-                en (bool):which sense resistor voltage is used; true will propably destroy your tmc
-        """
-        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
-        if en:
-            self.tmc_logger.log("activated internal sense resistors.",
-                                Loglevel.INFO.value)
-            self.tmc_logger.log("VREF pin internally is driven to GND in this mode.",
-                                Loglevel.INFO.value)
-            self.tmc_logger.log("This will most likely destroy your driver!!!",
-                                Loglevel.INFO.value)
-            raise SystemExit
-            # gconf = self.tmc_uart.set_bit(gconf, tmc_reg.internal_rsense)
-        self.tmc_logger.log("activated operation with external sense resistors",
-                            Loglevel.INFO.value)
-        gconf = self.tmc_uart.clear_bit(gconf, tmc_reg.internal_rsense)
-        self.tmc_uart.write_reg_check(tmc_reg.GCONF, gconf)
-
-
-
-    def set_irun_ihold(self, ihold, irun, ihold_delay):
-        """
-        sets the current scale (CS) for Running and Holding
-        and the delay, when to be switched to Holding current
-
-            Parameters:
-                ihold (int): 0-31
-                irun (int): 0-31
-                ihold_delay (int): 0-15
-        """
-        ihold_irun = 0
-
-        ihold_irun = ihold_irun | ihold << 0
-        ihold_irun = ihold_irun | irun << 8
-        ihold_irun = ihold_irun | ihold_delay << 16
-        self.tmc_logger.log("ihold_irun", Loglevel.INFO.value)
-        self.tmc_logger.log(str(bin(ihold_irun)), Loglevel.INFO.value)
-
-        self.tmc_logger.log("writing ihold_irun", Loglevel.INFO.value)
-        self.tmc_uart.write_reg_check(tmc_reg.IHOLD_IRUN, ihold_irun)
-
-
-
-    def set_pdn_disable(self,pdn_disable):
-        """
-        disables PDN on the UART pin
-        0: PDN_UART controls standstill current reduction
-        1: PDN_UART input function disabled. Set this bit,
-        when using the UART interface!
-
-            Parameters:
-                pdn_disable (bool): whether pdn should be disabled
-        """
-        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
-        if pdn_disable:
-            self.tmc_logger.log("enabled PDN_UART", Loglevel.INFO.value)
-            gconf = self.tmc_uart.set_bit(gconf, tmc_reg.pdn_disable)
-        else:
-            self.tmc_logger.log("disabled PDN_UART", Loglevel.INFO.value)
-            gconf = self.tmc_uart.clear_bit(gconf, tmc_reg.pdn_disable)
-        self.tmc_uart.write_reg_check(tmc_reg.GCONF, gconf)
-
-
-
-    def set_current(self, run_current, hold_current_multiplier = 0.5, hold_current_delay = 10,
-                    use_vref=False, vref = 1.2, pdn_disable = True):
-        """
-        sets the current flow for the motor. 
-
-            Parameters:
-                run_current (int): current during movement in mA
-                hold_current_multiplier (int): current multiplier during standstill
-                hold_current_delay (int): delay after standstill after which the current drops
-                use_vref (int): most driver have a trimpotentiometer to set the vref. optional
-                Vref (int): if you want to use Vref, you need to set the Vref Voltage
-                pdn_disable (bool): should be disabled if UART is used
-        """
-        cs_irun = 0
-        rsense = 0.11
-        vfs = 0
-
-        if use_vref:
-            self.set_iscale_analog(True)
-        else:
-            vref=2.5
-            self.set_iscale_analog(False)
-
-        vfs = 0.325 * vref / 2.5
-        cs_irun = 32.0*1.41421*run_current/1000.0*(rsense+0.02)/vfs - 1
-
-        # If Current Scale is too low, turn on high sensitivity VSsense and calculate again
-        if cs_irun < 16:
-            self.tmc_logger.log("CS too low; switching to VSense True", Loglevel.INFO.value)
-            vfs = 0.180 * vref / 2.5
-            cs_irun = 32.0*1.41421*run_current/1000.0*(rsense+0.02)/vfs - 1
-            self.setvsense(True)
-        else: # If CS >= 16, turn off high_senser
-            self.tmc_logger.log("CS in range; using VSense False", Loglevel.INFO.value)
-            self.setvsense(False)
-
-        cs_irun = min(cs_irun, 31)
-        cs_irun = max(cs_irun, 0)
-
-        CS_IHold = hold_current_multiplier * cs_irun
-
-        cs_irun = round(cs_irun)
-        CS_IHold = round(CS_IHold)
-        hold_current_delay = round(hold_current_delay)
-
-        self.tmc_logger.log("cs_irun: " + str(cs_irun), Loglevel.INFO.value)
-        self.tmc_logger.log("CS_IHold: " + str(CS_IHold), Loglevel.INFO.value)
-        self.tmc_logger.log("Delay: " + str(hold_current_delay), Loglevel.INFO.value)
-
-        # return (float)(CS+1)/32.0 * (vsense() ? 0.180 : 0.325)/(rsense+0.02) / 1.41421 * 1000;
-        run_current_actual = (cs_irun+1)/32.0 * (vfs)/(rsense+0.02) / 1.41421 * 1000
-        self.tmc_logger.log("actual current: "+str(round(run_current_actual))+" mA",
-                            Loglevel.INFO.value)
-
-        self.set_irun_ihold(CS_IHold, cs_irun, hold_current_delay)
-
-        self.set_pdn_disable(pdn_disable)
-
-
-
-    def get_spreadcycle(self):
-        """
-        return whether spreadcycle (1) is active or stealthchop (0)
-
-            Parameters:
-                en_spread (bool): true = spreadcycle; false = stealthchop
-        """
-        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
-        return gconf & tmc_reg.en_spreadcycle
-
-
-
-    def set_spreadcycle(self,en_spread):
-        """
-        enables spreadcycle (1) or stealthchop (0)
-
-            Parameters:
-                en_spread (bool): true to enable spreadcycle; false to enable stealthchop
-        """
-        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
-        if en_spread:
-            self.tmc_logger.log("activated Spreadcycle", Loglevel.INFO.value)
-            gconf = self.tmc_uart.set_bit(gconf, tmc_reg.en_spreadcycle)
-        else:
-            self.tmc_logger.log("activated Stealthchop", Loglevel.INFO.value)
-            gconf = self.tmc_uart.clear_bit(gconf, tmc_reg.en_spreadcycle)
-        self.tmc_uart.write_reg_check(tmc_reg.GCONF, gconf)
-
-
-
-    def get_interpolation(self):
-        """
-        return whether the tmc inbuilt interpolation is active
-
-            Returns:
-                en (bool): true if internal µstep interpolation is enabled
-        """
-        chopconf = self.tmc_uart.read_int(tmc_reg.CHOPCONF)
-        return bool(chopconf & tmc_reg.intpol)
-
-
-
-    def set_interpolation(self, en):
-        """
-        enables the tmc inbuilt interpolation of the steps to 256 microsteps
-
-            Parameters:
-                en (bool): true to enable internal µstep interpolation
-        """
-        chopconf = self.tmc_uart.read_int(tmc_reg.CHOPCONF)
-
-        if en:
-            chopconf = self.tmc_uart.set_bit(chopconf, tmc_reg.intpol)
-        else:
-            chopconf = self.tmc_uart.clear_bit(chopconf, tmc_reg.intpol)
-
-        self.tmc_logger.log("writing microstep interpolation setting: "+str(en),
-                            Loglevel.INFO.value)
-        self.tmc_uart.write_reg_check(tmc_reg.CHOPCONF, chopconf)
-
-
-
-    def get_microstepping_resolution(self):
-        """
-        returns the current native microstep resolution (1-256)
-
-            Returns:
-                msres (int): µstep resolution
-        """
-        chopconf = self.tmc_uart.read_int(tmc_reg.CHOPCONF)
-
-        msresdezimal = chopconf & (tmc_reg.msres0 | tmc_reg.msres1 |
-                                   tmc_reg.msres2 | tmc_reg.msres3)
-
-        msresdezimal = msresdezimal >> 24
-        msresdezimal = 8 - msresdezimal
-
-        self._msres = int(math.pow(2, msresdezimal))
-
-        return self._msres
-
-
-
-    def set_microstepping_resolution(self, msres):
-        """
-        sets the current native microstep resolution (1,2,4,8,16,32,64,128,256)
-
-            Parameters:
-                msres (int): µstep resolution; has to be a power of 2 or 1 for fullstep
-        """
-        chopconf = self.tmc_uart.read_int(tmc_reg.CHOPCONF)
-        #setting all bits to zero
-        chopconf = chopconf & (~tmc_reg.msres0 | ~tmc_reg.msres1 |
-                               ~tmc_reg.msres2 | ~tmc_reg.msres3)
-        msresdezimal = int(math.log(msres, 2))
-        msresdezimal = 8 - msresdezimal
-        chopconf = int(chopconf) & int(4043309055)
-        chopconf = chopconf | msresdezimal <<24
-
-        self.tmc_logger.log("writing "+str(msres)+" microstep setting", Loglevel.INFO.value)
-        self.tmc_uart.write_reg_check(tmc_reg.CHOPCONF, chopconf)
-
-        self.set_mstep_resolution_reg_select(True)
-
-        self.read_steps_per_rev()
-
-        return True
-
-
-
-    def set_mstep_resolution_reg_select(self, en):
-        """
-        sets the register bit "mstep_reg_select" to 1 or 0 depending to the given value.
-        this is needed to set the microstep resolution via UART
-        this method is called by "set_microstepping_resolution"
-
-            Parameters:
-                en (bool): true to set µstep resolution via UART
-        """
-        gconf = self.tmc_uart.read_int(tmc_reg.GCONF)
-
-        if en is True:
-            gconf = self.tmc_uart.set_bit(gconf, tmc_reg.mstep_reg_select)
-        else:
-            gconf = self.tmc_uart.clear_bit(gconf, tmc_reg.mstep_reg_select)
-
-        self.tmc_logger.log("writing MStep Reg Select: "+str(en), Loglevel.INFO.value)
-        self.tmc_uart.write_reg_check(tmc_reg.GCONF, gconf)
-
-
-
     def read_steps_per_rev(self):
         """
         returns how many steps are needed for one revolution.
@@ -989,35 +396,8 @@ class TMC_2209:
 
 
 
-    def get_interface_transmission_counter(self):
-        """
-        reads the interface transmission counter from the tmc register
-        this value is increased on every succesfull write access
-        can be used to verify a write access
-
-            Returns:
-                ifcnt (int): 8bit IFCNT Register
-        """
-        ifcnt = self.tmc_uart.read_int(tmc_reg.IFCNT)
-        self.tmc_logger.log("Interface Transmission Counter: "+str(ifcnt), Loglevel.INFO.value)
-        return ifcnt
-
-
-
-    def get_tstep(self):
-        """
-        reads the current tstep from the driver register
-
-            Returns:
-                tstep (int): TStep time
-        """
-        tstep = self.tmc_uart.read_int(tmc_reg.TSTEP)
-        return tstep
-
-
-
-    def set_vactual(self, vactual, duration=0, acceleration=0, show_stallguard_result=False,
-                    show_tstep=False):
+    def set_vactual_dur(self, vactual, duration=0, acceleration=0,
+                             show_stallguard_result=False, show_tstep=False):
         """
         sets the register bit "VACTUAL" to to a given value
         VACTUAL allows moving the motor by UART control.
@@ -1042,12 +422,12 @@ class TMC_2209:
 
         if duration != 0:
             self.tmc_logger.log("vactual: "+str(vactual)+" for "+str(duration)+" sec",
-                                Loglevel.INFO.value)
+                                Loglevel.INFO)
         else:
-            self.tmc_logger.log("vactual: "+str(vactual), Loglevel.INFO.value)
-        self.tmc_logger.log(str(bin(vactual)), Loglevel.INFO.value)
+            self.tmc_logger.log("vactual: "+str(vactual), Loglevel.INFO)
+        self.tmc_logger.log(str(bin(vactual)), Loglevel.INFO)
 
-        self.tmc_logger.log("writing vactual", Loglevel.INFO.value)
+        self.tmc_logger.log("writing vactual", Loglevel.INFO)
         if acceleration == 0:
             self.tmc_uart.write_reg_check(tmc_reg.VACTUAL, vactual)
 
@@ -1065,18 +445,21 @@ class TMC_2209:
                     time_to_stop = current_time-1
             if acceleration != 0 and current_time > time_to_stop:
                 current_vactual -= acceleration*sleeptime
-                self.tmc_uart.write_reg_check(tmc_reg.VACTUAL, int(round(current_vactual)))
+                self.tmc_uart.write_reg_check(tmc_reg.VACTUAL,
+                                                       int(round(current_vactual)))
                 time.sleep(sleeptime)
             elif acceleration != 0 and abs(current_vactual)<abs(vactual):
                 current_vactual += acceleration*sleeptime
-                self.tmc_uart.write_reg_check(tmc_reg.VACTUAL, int(round(current_vactual)))
+                self.tmc_uart.write_reg_check(tmc_reg.VACTUAL,
+                                                       int(round(current_vactual)))
                 time.sleep(sleeptime)
             if show_stallguard_result:
-                self.tmc_logger.log("StallGuard result: "+str(self.get_stallguard_result()),
-                            Loglevel.INFO.value)
+                self.tmc_logger.log("StallGuard result: "+
+                                    str(self.get_stallguard_result()), Loglevel.INFO)
                 time.sleep(0.1)
             if show_tstep:
-                self.tmc_logger.log("TStep result: "+str(self.get_tstep()), Loglevel.INFO.value)
+                self.tmc_logger.log("TStep result: "+str(self.get_tstep()),
+                                    Loglevel.INFO)
                 time.sleep(0.1)
             current_time = time.time()
         self.tmc_uart.write_reg_check(tmc_reg.VACTUAL, 0)
@@ -1107,7 +490,7 @@ class TMC_2209:
         duration = abs(revolutions/rps)
         if revolutions<0:
             vactual = -vactual
-        return self.set_vactual(vactual, duration, acceleration=acceleration)
+        return self.set_vactual_dur(vactual, duration, acceleration=acceleration)
 
 
 
@@ -1129,56 +512,6 @@ class TMC_2209:
 
 
 
-    def get_stallguard_result(self):
-        """
-        return the current stallguard result
-        its will be calculated with every fullstep
-        higher values means a lower motor load
-
-            Returns:
-                sg_result (int): StallGuard Result
-        """
-        sg_result = self.tmc_uart.read_int(tmc_reg.SG_RESULT)
-        return sg_result
-
-
-
-    def set_stallguard_threshold(self, threshold):
-        """
-        sets the register bit "SGTHRS" to to a given value
-        this is needed for the stallguard interrupt callback
-        SG_RESULT becomes compared to the double of this threshold.
-        SG_RESULT ≤ SGTHRS*2
-
-            Parameters:
-                threshold (int): value for SGTHRS
-        """
-
-        self.tmc_logger.log("sgthrs", Loglevel.INFO.value)
-        self.tmc_logger.log(str(bin(threshold)), Loglevel.INFO.value)
-
-        self.tmc_logger.log("writing sgthrs", Loglevel.INFO.value)
-        self.tmc_uart.write_reg_check(tmc_reg.SGTHRS, threshold)
-
-
-
-    def set_coolstep_threshold(self, threshold):
-        """
-        This  is  the  lower  threshold  velocity  for  switching  
-        on  smart energy CoolStep and StallGuard to DIAG output. (unsigned)
-
-            Parameters:
-                threshold (int): threshold  velocity
-        """
-
-        self.tmc_logger.log("tcoolthrs", Loglevel.INFO.value)
-        self.tmc_logger.log(str(bin(threshold)), Loglevel.INFO.value)
-
-        self.tmc_logger.log("writing tcoolthrs", Loglevel.INFO.value)
-        self.tmc_uart.write_reg_check(tmc_reg.TCOOLTHRS, threshold)
-
-
-
     def set_stallguard_callback(self, pin_stallguard, threshold, callback,
                                 min_speed = 100, ignore_delay = 0):
         """
@@ -1193,12 +526,13 @@ class TMC_2209:
                 min_speed (int): min speed [steps/s] for StallGuard
         """
         self.tmc_logger.log("setup stallguard callback on GPIO"+str(pin_stallguard),
-                            Loglevel.INFO.value)
+                            Loglevel.INFO)
         self.tmc_logger.log("StallGuard Threshold: "+str(threshold)+"\tminimum Speed: "+
-                            str(min_speed), Loglevel.INFO.value)
+                            str(min_speed), Loglevel.INFO)
 
         self.set_stallguard_threshold(threshold)
-        self.set_coolstep_threshold(tmc_math.steps_to_tstep(min_speed, self._msres))
+        self.set_coolstep_threshold(tmc_math.steps_to_tstep(
+            min_speed, self.get_microstepping_resolution()))
         self._sg_delay = ignore_delay
         self._sg_callback = callback
         self._pin_stallguard = pin_stallguard
@@ -1221,395 +555,11 @@ class TMC_2209:
         """
         del gpio_pin
         if self._sg_callback is None:
-            self.tmc_logger.log("StallGuard callback is None", Loglevel.DEBUG.value)
+            self.tmc_logger.log("StallGuard callback is None", Loglevel.DEBUG)
             return
         if time.time()<=self._starttime+self._sg_delay and self._sg_delay != 0:
             return
         self._sg_callback()
-
-
-
-    def get_microstep_counter(self):
-        """
-        returns the current Microstep counter.
-        Indicates actual position in the microstep table for CUR_A
-
-            Returns:
-                mscnt (int): current Microstep counter
-        """
-        mscnt = self.tmc_uart.read_int(tmc_reg.MSCNT)
-        return mscnt
-
-
-
-    def get_microstep_counter_in_steps(self, offset=0):
-        """
-        returns the current Microstep counter.
-        Indicates actual position in the microstep table for CUR_A
-
-            Parameters:
-                offset (int): offset added to the current step counter
-
-            Returns:
-                step (int): current Microstep counter convertet to steps
-        """
-        step = (self.get_microstep_counter()-64)*(self._msres*4)/1024
-        step = (4*self._msres)-step-1
-        step = round(step)
-        return step+offset
-
-
-
-    def set_max_speed(self, speed):
-        """
-        sets the maximum motor speed in µsteps per second
-
-            Parameters:
-                speed (int): maximum speed in microsteps/sec
-        """
-        if speed < 0.0:
-            speed = -speed
-        if self._max_speed != speed:
-            self._max_speed = speed
-            self._cmin = 1000000.0 / speed
-            # Recompute _n from current speed and adjust speed if accelerating or cruising
-            if self._n > 0:
-                self._n = (self._speed * self._speed) / (2.0 * self._acceleration) # Equation 16
-                self.compute_new_speed()
-
-
-
-    def set_max_speed_fullstep(self, speed):
-        """
-        sets the maximum motor speed in fullsteps per second
-
-            Parameters:
-                speed (int): maximum speed in fullsteps/sec
-        """
-        self.set_max_speed(speed*self._msres)
-
-
-
-    def get_max_speed(self):
-        """
-        returns the maximum motor speed in steps per second
-
-            Returns:
-                max_speed (int): current maximum speed in steps/sec
-        """
-        return self._max_speed
-
-
-
-    def set_acceleration(self, acceleration):
-        """
-        sets the motor acceleration/decceleration in µsteps per sec per sec
-
-            Parameters:
-                acceleration (int): acceleration/decceleration in µsteps per sec per sec
-        """
-        if acceleration == 0.0:
-            return
-        acceleration = abs(acceleration)
-        if self._acceleration != acceleration:
-            # Recompute _n per Equation 17
-            self._n = self._n * (self._acceleration / acceleration)
-            # New c0 per Equation 7, with correction per Equation 15
-            self._c0 = 0.676 * math.sqrt(2.0 / acceleration) * 1000000.0 # Equation 15
-            self._acceleration = acceleration
-            self.compute_new_speed()
-
-
-
-    def set_acceleration_fullstep(self, acceleration):
-        """
-        sets the motor acceleration/decceleration in fullsteps per sec per sec
-
-            Parameters:
-                acceleration (int): acceleration/decceleration in fullsteps per sec per sec
-        """
-        self.set_acceleration(acceleration*self._msres)
-
-
-
-    def get_acceleration(self):
-        """
-        returns the motor acceleration/decceleration in steps per sec per sec
-
-            Returns:
-                acceleration (int): acceleration/decceleration in µsteps per sec per sec
-        """
-        return self._acceleration
-
-
-
-    def stop(self, stop_mode = StopMode.HARDSTOP):
-        """
-        stop the current movement
-
-            Parameters:
-                stop_mode (enum): whether the movement should be stopped immediatly or softly
-        """
-        self._stop = stop_mode
-
-
-
-    def get_movement_phase(self):
-        """
-        return the current Movement Phase
-
-            Returns:
-                movement_phase (enum): current Movement Phase
-        """
-        return self._movement_phase
-
-
-
-    def run_to_position_steps(self, steps, movement_abs_rel = None):
-        """
-        runs the motor to the given position.
-        with acceleration and deceleration
-        blocks the code until finished or stopped from a different thread!
-        returns true when the movement if finshed normally and false,
-        when the movement was stopped
-
-            Parameters:
-                steps (int): amount of steps; can be negative
-                movement_abs_rel (enum): whether the movement should be absolut or relative
-
-            Returns:
-                stop (enum): how the movement was finished
-        """
-        if movement_abs_rel is None:
-            movement_abs_rel = self._movement_abs_rel
-
-        if movement_abs_rel == MovementAbsRel.RELATIVE:
-            self._target_pos = self._current_pos + steps
-        else:
-            self._target_pos = steps
-
-        self._stop = StopMode.NO
-        self._step_interval = 0
-        self._speed = 0.0
-        self._n = 0
-        self.compute_new_speed()
-        while self.run(): #returns false, when target position is reached
-            if self._stop == StopMode.HARDSTOP:
-                break
-
-        self._movement_phase = MovementPhase.STANDSTILL
-        return self._stop
-
-
-
-    def run_to_position_revolutions(self, revolutions, movement_absolute_relative = None):
-        """
-        runs the motor to the given position.
-        with acceleration and deceleration
-        blocks the code until finished!
-
-            Parameters:
-                revolutions (int): amount of revs; can be negative
-                movement_abs_rel (enum): whether the movement should be absolut or relative
-
-            Returns:
-                stop (enum): how the movement was finished
-        """
-        return self.run_to_position_steps(round(revolutions * self._steps_per_rev),
-                                          movement_absolute_relative)
-
-
-
-    def run_to_position_steps_threaded(self, steps, movement_abs_rel = None):
-        """
-        runs the motor to the given position.
-        with acceleration and deceleration
-        does not block the code
-        returns true when the movement if finshed normally and false,
-        when the movement was stopped
-
-            Parameters:
-                steps (int): amount of steps; can be negative
-                movement_abs_rel (enum): whether the movement should be absolut or relative
-
-            Returns:
-                stop (enum): how the movement was finished
-        """
-        self._movement_thread = threading.Thread(target=self.run_to_position_steps,
-                                                 args=(steps, movement_abs_rel))
-        self._movement_thread.start()
-
-
-
-    def run_to_position_revolutions_threaded(self, revolutions, movement_absolute_relative = None):
-        """
-        runs the motor to the given position.
-        with acceleration and deceleration
-        does not block the code
-
-            Parameters:
-                revolutions (int): amount of revs; can be negative
-                movement_abs_rel (enum): whether the movement should be absolut or relative
-
-            Returns:
-                stop (enum): how the movement was finished
-        """
-        return self.run_to_position_steps_threaded(round(revolutions * self._steps_per_rev),
-                                                   movement_absolute_relative)
-
-
-
-    def wait_for_movement_finished_threaded(self):
-        """
-        wait for the motor to finish the movement,
-        if startet threaded
-        returns true when the movement if finshed normally and false,
-        when the movement was stopped
-
-            Returns:
-                stop (enum): how the movement was finished
-        """
-        self._movement_thread.join()
-        return self._stop
-
-
-
-    def run(self):
-        """
-        calculates a new speed if a speed was made
-        returns true if the target position is reached
-        should not be called from outside!
-        """
-        if self.run_speed(): #returns true, when a step is made
-            self.compute_new_speed()
-            #self.tmc_logger.log(self.get_stallguard_result())
-            #self.tmc_logger.log(self.get_tstep())
-        return (self._speed != 0.0 and self.distance_to_go() != 0)
-
-
-
-    def distance_to_go(self):
-        """
-        returns the remaining distance the motor should run
-        """
-        return self._target_pos - self._current_pos
-
-
-
-    def compute_new_speed(self):
-        """
-        returns the calculated current speed depending on the acceleration
-        this code is based on: 
-        "Generate stepper-motor speed profiles in real time" by David Austin
-
-        https://www.embedded.com/generate-stepper-motor-speed-profiles-in-real-time/
-        https://web.archive.org/web/20140705143928/http://fab.cba.mit.edu/classes/MIT/961.09/projects/i0/Stepper_Motor_Speed_Profile.pdf
-        """
-        distance_to = self.distance_to_go() # +ve is clockwise from curent location
-        steps_to_stop = (self._speed * self._speed) / (2.0 * self._acceleration) # Equation 16
-        if ((distance_to == 0 and steps_to_stop <= 2) or
-        (self._stop == StopMode.SOFTSTOP and steps_to_stop <= 1)):
-            # We are at the target and its time to stop
-            self._step_interval = 0
-            self._speed = 0.0
-            self._n = 0
-            self._movement_phase = MovementPhase.STANDSTILL
-            self.tmc_logger.log("time to stop", Loglevel.MOVEMENT.value)
-            return
-
-        if distance_to > 0:
-            # We are anticlockwise from the target
-            # Need to go clockwise from here, maybe decelerate now
-            if self._n > 0:
-                # Currently accelerating, need to decel now? Or maybe going the wrong way?
-                if ((steps_to_stop >= distance_to) or self._direction == Direction.CCW or
-                    self._stop == StopMode.SOFTSTOP):
-                    self._n = -steps_to_stop # Start deceleration
-                    self._movement_phase = MovementPhase.DECELERATING
-            elif self._n < 0:
-                # Currently decelerating, need to accel again?
-                if (steps_to_stop < distance_to) and self._direction == Direction.CW:
-                    self._n = -self._n # Start accceleration
-                    self._movement_phase = MovementPhase.ACCELERATING
-        elif distance_to < 0:
-            # We are clockwise from the target
-            # Need to go anticlockwise from here, maybe decelerate
-            if self._n > 0:
-                # Currently accelerating, need to decel now? Or maybe going the wrong way?
-                if (((steps_to_stop >= -distance_to) or self._direction == Direction.CW or
-                    self._stop == StopMode.SOFTSTOP)):
-                    self._n = -steps_to_stop # Start deceleration
-                    self._movement_phase = MovementPhase.DECELERATING
-            elif self._n < 0:
-                # Currently decelerating, need to accel again?
-                if (steps_to_stop < -distance_to) and self._direction == Direction.CCW:
-                    self._n = -self._n # Start accceleration
-                    self._movement_phase = MovementPhase.ACCELERATING
-        # Need to accelerate or decelerate
-        if self._n == 0:
-            # First step from stopped
-            self._cn = self._c0
-            GPIO.output(self._pin_step, GPIO.LOW)
-            #self.tmc_logger.log("distance to: " + str(distance_to))
-            if distance_to > 0:
-                self.set_direction_pin(1)
-                self.tmc_logger.log("going CW", Loglevel.MOVEMENT.value)
-            else:
-                self.set_direction_pin(0)
-                self.tmc_logger.log("going CCW", Loglevel.MOVEMENT.value)
-            self._movement_phase = MovementPhase.ACCELERATING
-        else:
-            # Subsequent step. Works for accel (n is +_ve) and decel (n is -ve).
-            self._cn = self._cn - ((2.0 * self._cn) / ((4.0 * self._n) + 1)) # Equation 13
-            self._cn = max(self._cn, self._cmin)
-            if self._cn == self._cmin:
-                self._movement_phase = MovementPhase.MAXSPEED
-        self._n += 1
-        self._step_interval = self._cn
-        self._speed = 1000000.0 / self._cn
-        if self._direction == 0:
-            self._speed = -self._speed
-
-
-
-    def run_speed(self):
-        """
-        this methods does the actual steps with the current speed
-        """
-        # Dont do anything unless we actually have a step interval
-        if not self._step_interval:
-            return False
-
-        curtime = time.time_ns()/1000
-
-        #self.tmc_logger.log("current time: " + str(curtime))
-        #self.tmc_logger.log("last st time: " + str(self._last_step_time))
-
-        if curtime - self._last_step_time >= self._step_interval:
-
-            if self._direction == 1: # Clockwise
-                self._current_pos += 1
-            else: # Anticlockwise
-                self._current_pos -= 1
-            self.make_a_step()
-
-            self._last_step_time = curtime # Caution: does not account for costs in step()
-            return True
-        return False
-
-
-
-    def make_a_step(self):
-        """
-        method that makes on step
-        for the TMC2209 there needs to be a signal duration of minimum 100 ns
-        """
-        GPIO.output(self._pin_step, GPIO.HIGH)
-        time.sleep(1/1000/1000)
-        GPIO.output(self._pin_step, GPIO.LOW)
-        time.sleep(1/1000/1000)
-
-        self.tmc_logger.log("one step", Loglevel.MOVEMENT.value)
 
 
 
@@ -1625,7 +575,7 @@ class TMC_2209:
         GPIO.output(self._pin_dir, GPIO.HIGH)
         GPIO.output(self._pin_en, GPIO.HIGH)
         time.sleep(0.1)
-        ioin = self.readIOIN()
+        ioin = self.read_ioin()
         if not ioin & tmc_reg.io_dir:
             pin_dir_ok = False
         if not ioin & tmc_reg.io_step:
@@ -1637,7 +587,7 @@ class TMC_2209:
         GPIO.output(self._pin_dir, GPIO.LOW)
         GPIO.output(self._pin_en, GPIO.LOW)
         time.sleep(0.1)
-        ioin = self.readIOIN()
+        ioin = self.read_ioin()
         if ioin & tmc_reg.io_dir:
             pin_dir_ok = False
         if ioin & tmc_reg.io_step:
@@ -1649,7 +599,7 @@ class TMC_2209:
         GPIO.output(self._pin_dir, GPIO.HIGH)
         GPIO.output(self._pin_en, GPIO.HIGH)
         time.sleep(0.1)
-        ioin = self.readIOIN()
+        ioin = self.read_ioin()
         if not ioin & tmc_reg.io_dir:
             pin_dir_ok = False
         if not ioin & tmc_reg.io_step:
@@ -1702,37 +652,37 @@ class TMC_2209:
         snd = result[0]
         rtn = result[1]
 
-        self.tmc_logger.log("length snd: "+str(len(snd)), Loglevel.DEBUG.value)
-        self.tmc_logger.log("length rtn: "+str(len(rtn)), Loglevel.DEBUG.value)
+        self.tmc_logger.log("length snd: "+str(len(snd)), Loglevel.DEBUG)
+        self.tmc_logger.log("length rtn: "+str(len(rtn)), Loglevel.DEBUG)
 
         if len(rtn)==12:
             self.tmc_logger.log("""the Raspberry Pi received the sended
-                                bits and the answer from the TMC""",Loglevel.INFO.value)
+                                bits and the answer from the TMC""",Loglevel.INFO)
         elif len(rtn)==4:
             self.tmc_logger.log("the Raspberry Pi received only the sended bits",
-                                Loglevel.INFO.value)
+                                Loglevel.INFO)
         elif len(rtn)==0:
             self.tmc_logger.log("the Raspberry Pi did not receive anything",
-                                Loglevel.INFO.value)
+                                Loglevel.INFO)
         else:
             self.tmc_logger.log("the Raspberry Pi received an unexpected amount of bits: "+
-                     str(len(rtn)), Loglevel.INFO.value)
+                     str(len(rtn)), Loglevel.INFO)
 
         if snd[0:4] == rtn[0:4]:
             self.tmc_logger.log("""the Raspberry Pi received exactly the bits it has send.
-                     the first 4 bits are the same""", Loglevel.INFO.value)
+                     the first 4 bits are the same""", Loglevel.INFO)
         else:
             self.tmc_logger.log("""the Raspberry Pi did not received the bits it has send.
-                     the first 4 bits are different""", Loglevel.INFO.value)
+                     the first 4 bits are different""", Loglevel.INFO)
 
 
-        self.tmc_logger.log("complete", Loglevel.DEBUG.value)
-        self.tmc_logger.log(str(snd.hex()), Loglevel.DEBUG.value)
-        self.tmc_logger.log(str(rtn.hex()), Loglevel.DEBUG.value)
+        self.tmc_logger.log("complete", Loglevel.DEBUG)
+        self.tmc_logger.log(str(snd.hex()), Loglevel.DEBUG)
+        self.tmc_logger.log(str(rtn.hex()), Loglevel.DEBUG)
 
-        self.tmc_logger.log("just the first 4 bits", Loglevel.DEBUG.value)
-        self.tmc_logger.log(str(snd[0:4].hex()), Loglevel.DEBUG.value)
-        self.tmc_logger.log(str(rtn[0:4].hex()), Loglevel.DEBUG.value)
+        self.tmc_logger.log("just the first 4 bits", Loglevel.DEBUG)
+        self.tmc_logger.log(str(snd[0:4].hex()), Loglevel.DEBUG)
+        self.tmc_logger.log(str(rtn[0:4].hex()), Loglevel.DEBUG)
 
 
         self.tmc_logger.log("---")
@@ -1750,8 +700,8 @@ class TMC_2209:
                 steps (int): amount of steps; can be negative
         """
 
-        self.tmc_logger.log("---", Loglevel.INFO.value)
-        self.tmc_logger.log("test_stallguard_threshold", Loglevel.INFO.value)
+        self.tmc_logger.log("---", Loglevel.INFO)
+        self.tmc_logger.log("test_stallguard_threshold", Loglevel.INFO)
 
         self.set_spreadcycle(0)
 
@@ -1766,7 +716,7 @@ class TMC_2209:
             stallguard_result = self.get_stallguard_result()
 
             self.tmc_logger.log(str(self._movement_phase) + " | " + str(stallguard_result),
-                     Loglevel.INFO.value)
+                     Loglevel.INFO)
 
             if (self._movement_phase == MovementPhase.ACCELERATING and
                stallguard_result < min_stallguard_result_accel):
@@ -1780,11 +730,11 @@ class TMC_2209:
 
         self.wait_for_movement_finished_threaded()
 
-        self.tmc_logger.log("---", Loglevel.INFO.value)
+        self.tmc_logger.log("---", Loglevel.INFO)
         self.tmc_logger.log("min StallGuard result during acceleration: " +
-                 str(min_stallguard_result_accel), Loglevel.INFO.value)
+                 str(min_stallguard_result_accel), Loglevel.INFO)
         self.tmc_logger.log("min StallGuard result during maxspeed: " +
-                 str(min_stallguard_result_maxspeed), Loglevel.INFO.value)
+                 str(min_stallguard_result_maxspeed), Loglevel.INFO)
         self.tmc_logger.log("min StallGuard result during deceleration: " +
-                 str(min_stallguard_result_decel), Loglevel.INFO.value)
-        self.tmc_logger.log("---", Loglevel.INFO.value)
+                 str(min_stallguard_result_decel), Loglevel.INFO)
+        self.tmc_logger.log("---", Loglevel.INFO)
