+++ tests/test_script_06_multiple_drivers.py
@@ -49,7 +49,7 @@ tmc2.set_movement_abs_rel(MovementAbsRel.ABSOLUTE)
 print("---")
 print("IOIN tmc1")
 print("---")
-tmc1.readIOIN()
+tmc1.read_ioin()
 
 print("---\n---")
 
@@ -57,7 +57,7 @@ print("---\n---")
 print("---")
 print("IOIN tmc2")
 print("---")
-tmc2.readIOIN()
+tmc2.read_ioin()
 
 print("---\n---")
